refactor: move viewParams to shared
