+++ src/components/BarChart.tsx
@@ -8,7 +8,7 @@ import { dateFormatShort, expandIntervalToRange } from "~/shared/util"
 import { useSelectedCategories } from "~/state/stores/selection"
 import { useClickedObject, useBlobColor, useBlobColors } from "~/state/stores/clicked-object"
 import { useQueryStates } from "nuqs"
-import { viewSearchParamsConfig } from "~/routes/viewParams"
+import { viewSearchParamsConfig } from "~/shared/viewParams"
 import { useOptions } from "~/contexts/OptionsContext"
 import { useMetrics } from "~/contexts/MetricContext"
 import { useGradient } from "~/hooks/svg"
