+++ src/components/Breadcrumb.tsx
@@ -4,7 +4,7 @@ import { useMemo, Fragment } from "react"
 import { cn } from "~/styling"
 import { useQueryStates } from "nuqs"
 import { href, Link } from "react-router"
-import { viewSearchParamsConfig } from "~/routes/viewParams"
+import { viewSearchParamsConfig } from "~/shared/viewParams"
 import { useDataNullable } from "~/contexts/DataContext"
 import { getSep } from "~/shared/util"
 import { AnalysisInfo } from "~/components/GlobalInfo"
