+++ src/components/inspection/ContributorsInspection.tsx
@@ -4,7 +4,7 @@ import { useFetcher, href } from "react-router"
 import { LegendDot } from "~/components/util"
 import { useData } from "~/contexts/DataContext"
 import { useMetrics } from "~/contexts/MetricContext"
-import { viewSearchParamsConfig, viewSerializer } from "~/routes/viewParams"
+import { viewSearchParamsConfig, viewSerializer } from "~/shared/viewParams"
 import type { loader } from "~/routes/api.contributor-distribution"
 import { useClickedObjectPath } from "~/state/stores/clicked-object"
 import { cn } from "~/styling"
