+++ src/components/inspection/MetricsInspection.tsx
@@ -38,7 +38,7 @@ import { LinesChangedMetric } from "~/metrics/linesChanged"
 import { LastChangedMetric } from "~/metrics/lastChanged"
 import { GroupContributorsModal } from "~/components/modals/GroupContributorsModal"
 import { useQueryState, useQueryStates } from "nuqs"
-import { viewSearchParamsConfig, viewSerializer } from "~/routes/viewParams"
+import { viewSearchParamsConfig, viewSerializer } from "~/shared/viewParams"
 import type { HexColor } from "~/shared/model"
 import { ZoomButtons } from "~/components/buttons/ZoomButtons"
 import type { loader } from "~/routes/api.metrics"
