+++ src/state/stores/clicked-object.ts
@@ -8,7 +8,7 @@ import { LastChangedMetric } from "~/metrics/lastChanged"
 import { LinesChangedMetric } from "~/metrics/linesChanged"
 import type { MetricCache } from "~/metrics/metrics"
 import { CommitsMetric } from "~/metrics/mostCommits"
-import { viewSearchParamsConfig } from "~/routes/viewParams"
+import { viewSearchParamsConfig } from "~/shared/viewParams"
 import type { ClickedObjectInfo, GitObjectNoChildren, HexColor, RawGitObject } from "~/shared/model"
 import { isTree } from "~/shared/util"
 import { useSelectedCategories, useIsCategorySelected } from "~/state/stores/selection"
